pyproject.toml
[tool.poetry]
name = "retrieve-research-paper"
version = "0.1.0"
description = ""
authors = ["h-k-nyosu <kuro.h1020@gmail.com>"]
readme = "README.md"
packages = [{include = "retrieve_research_paper"}]

[tool.poetry.dependencies]
python = "^3.10"
arxiv = "^1.4.7"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
pytz = "^2023.3"
asyncio = "^3.4.3"
openai = "^0.27.4"
slack-sdk = {extras = ["aiohttp"], version = "^3.21.2"}
slack-bolt = {extras = ["aiohttp"], version = "^1.17.2"}
APScheduler = "^3.10.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
